minor changes with dapp files

+++ contracts/dex/ERC20Events.sol
@@ -2,6 +2,6 @@
 pragma solidity ^0.4.18;
 
 contract ERC20Events {
-    event Approval(address indexed src, address indexed guy, uint wad);
-    event Transfer(address indexed src, address indexed dst, uint wad);
+    event Approval(address indexed src, address indexed guy, uint256 wad);
+    event Transfer(address indexed src, address indexed dst, uint256 wad);
 }
