+++ contracts/dex/MatchingEvents.sol
@@ -5,12 +5,12 @@ import {ERC20} from "./ERC20.sol";
 
 contract MatchingEvents {
     event LogBuyEnabled(bool isEnabled);
-    event LogMinSell(address pay_gem, uint min_amount);
+    event LogMinSell(address pay_gem, uint256 min_amount);
     event LogMatchingEnabled(bool isEnabled);
-    event LogUnsortedOffer(uint id);
-    event LogSortedOffer(uint id);
+    event LogUnsortedOffer(uint256 id);
+    event LogSortedOffer(uint256 id);
     event LogAddTokenPairWhitelist(ERC20 baseToken, ERC20 quoteToken);
     event LogRemTokenPairWhitelist(ERC20 baseToken, ERC20 quoteToken);
-    event LogInsert(address keeper, uint id);
-    event LogDelete(address keeper, uint id);
+    event LogInsert(address keeper, uint256 id);
+    event LogDelete(address keeper, uint256 id);
 }
